requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
Django==6.0.2
django-cors-headers==4.9.0
django-extensions==4.1
django-oauth-toolkit==3.2.0
djangorestframework==3.16.1
idna==3.11
jwcrypto==1.5.6
oauthlib==3.3.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
